# client/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/logged":{"get":{"summary":"Logged","operationId":"getLoggedUser","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/user/{userID}":{"get":{"summary":"Getuserbyid","operationId":"getUser","parameters":[{"name":"userID","in":"path","required":true,"schema":{"type":"integer","title":"Userid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/posts/{postID}":{"get":{"summary":"Getpostbyid","operationId":"getPost","parameters":[{"name":"postID","in":"path","required":true,"schema":{"type":"integer","title":"Postid"}},{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/posts":{"get":{"summary":"Getallposts","operationId":"getPosts","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"},"title":"Response Getposts"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/posts/user/{userID}":{"get":{"summary":"Getallpostsfromuser","operationId":"getUserPosts","parameters":[{"name":"userID","in":"path","required":true,"schema":{"type":"integer","title":"Userid"}},{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"},"title":"Response Getuserposts"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/comments/{postID}":{"get":{"summary":"Getallcommentsfrompost","operationId":"getPostComments","parameters":[{"name":"postID","in":"path","required":true,"schema":{"type":"integer","title":"Postid"}},{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommentResponse"},"title":"Response Getpostcomments"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/posts/create":{"post":{"summary":"Createnewpost","operationId":"createPost","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewPost"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/posts/delete/{postID}":{"delete":{"summary":"Deletepost","operationId":"deletePost","parameters":[{"name":"postID","in":"path","required":true,"schema":{"type":"integer","title":"Postid"}},{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/posts/comment":{"post":{"summary":"Createnewcomment","operationId":"createComment","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewComment"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/posts/like":{"post":{"summary":"Likepost","operationId":"likePost","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostRef"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Unlikepost","operationId":"unlikePost","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostRef"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/comments/like":{"post":{"summary":"Likecomment","operationId":"likeComment","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentRef"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Unlikecomment","operationId":"unlikeComment","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentRef"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/comments/best":{"put":{"summary":"Bestcomment","operationId":"chooseBestComment","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BestCommentRef"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/posts/view":{"post":{"summary":"Viewpost","operationId":"viewPost","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostRef"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/posts/closed":{"put":{"summary":"Closeopenpost","operationId":"togglePostClosed","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostRef"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/register":{"post":{"summary":"Createnewuser","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewUser"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/login":{"post":{"summary":"Login","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPayload"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/logout":{"post":{"summary":"Logout","operationId":"logout","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uid"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BestCommentRef":{"properties":{"post_id":{"type":"integer","title":"Post Id"},"comment_id":{"type":"integer","title":"Comment Id"}},"type":"object","required":["post_id","comment_id"],"title":"BestCommentRef"},"CommentRef":{"properties":{"comment_id":{"type":"integer","title":"Comment Id"}},"type":"object","required":["comment_id"],"title":"CommentRef"},"CommentResponse":{"properties":{"id":{"type":"integer","title":"Id"},"content":{"type":"string","title":"Content"},"post_id":{"type":"integer","title":"Post Id"},"user_id":{"type":"integer","title":"User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"user":{"anyOf":[{"$ref":"#/components/schemas/UserResponse"},{"type":"null"}]},"like_count":{"type":"integer","title":"Like Count"},"is_liked":{"type":"boolean","title":"Is Liked"}},"type":"object","required":["id","content","post_id","user_id","created_at","updated_at","like_count","is_liked"],"title":"CommentResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"NewComment":{"properties":{"post_id":{"type":"integer","title":"Post Id"},"content":{"type":"string","maxLength":10000,"minLength":1,"title":"Content"}},"type":"object","required":["post_id","content"],"title":"NewComment"},"NewPost":{"properties":{"title":{"type":"string","maxLength":120,"minLength":3,"title":"Title"},"content":{"type":"string","maxLength":10000,"minLength":1,"title":"Content"}},"type":"object","required":["title","content"],"title":"NewPost"},"NewUser":{"properties":{"username":{"type":"string","maxLength":30,"minLength":3,"pattern":"^[a-zA-Z0-9_]+$","title":"Username"},"nickname":{"type":"string","maxLength":30,"minLength":1,"title":"Nickname"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":72,"minLength":6,"title":"Password"}},"type":"object","required":["username","nickname","email","password"],"title":"NewUser"},"PostRef":{"properties":{"post_id":{"type":"integer","title":"Post Id"}},"type":"object","required":["post_id"],"title":"PostRef"},"PostResponse":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"content":{"type":"string","title":"Content"},"is_closed":{"type":"boolean","title":"Is Closed"},"answer_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Answer Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"user_id":{"type":"integer","title":"User Id"},"user":{"$ref":"#/components/schemas/UserResponse"},"like_count":{"type":"integer","title":"Like Count"},"view_count":{"type":"integer","title":"View Count"},"is_liked":{"type":"boolean","title":"Is Liked"}},"type":"object","required":["id","title","content","is_closed","answer_id","created_at","updated_at","user_id","user","like_count","view_count","is_liked"],"title":"PostResponse"},"UserPayload":{"properties":{"user":{"type":"string","title":"User"},"password":{"type":"string","title":"Password"}},"type":"object","required":["user","password"],"title":"UserPayload"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"nickname":{"type":"string","title":"Nickname"},"username":{"type":"string","title":"Username"},"email":{"type":"string","title":"Email"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","nickname","username","email","created_at","updated_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}